+++ batch/batch/front_end/front_end.py
@@ -18,7 +18,7 @@ from hailtop.utils import time_msecs, time_msecs_str, humanize_timedelta_msecs, 
     request_retry_transient_errors, run_if_changed, retry_long_running, \
     LoggingTimer
 from hailtop.config import get_deploy_config
-from hailtop.ssl import get_server_ssl_context, ssl_client_session
+from hailtop.tls import get_server_ssl_context, ssl_client_session
 from gear import Database, setup_aiohttp_session, \
     rest_authenticated_users_only, web_authenticated_users_only, \
     web_authenticated_developers_only, check_csrf_token, transaction, \
