+++ ci/test/test_ci.py
@@ -4,7 +4,7 @@ import asyncio
 
 from hailtop.config import get_deploy_config
 from hailtop.auth import service_auth_headers
-from hailtop.ssl import ssl_client_session
+from hailtop.tls import ssl_client_session
 import hailtop.utils as utils
 
 pytestmark = pytest.mark.asyncio
