+++ gear/gear/auth.py
@@ -6,7 +6,7 @@ from aiohttp import web
 import aiohttp_session
 from hailtop.config import get_deploy_config
 from hailtop.utils import request_retry_transient_errors
-from hailtop.ssl import ssl_client_session
+from hailtop.tls import ssl_client_session
 
 log = logging.getLogger('gear.auth')
