+++ hail/python/hailtop/batch_client/aioclient.py
@@ -11,7 +11,7 @@ from asyncinit import asyncinit
 from hailtop.config import get_deploy_config
 from hailtop.auth import async_get_userinfo, service_auth_headers
 from hailtop.utils import bounded_gather, request_retry_transient_errors, tqdm, TQDM_DEFAULT_DISABLE
-from hailtop.ssl import ssl_client_session
+from hailtop.tls import ssl_client_session
 
 from .globals import tasks, complete_states
